projects/HebaChain-contracts/pyproject.toml
[tool.poetry]
name = "HebaChain-contracts"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["balo <98645847+balojey@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
algokit-utils = "^2.3.0"
python-dotenv = "^1.0.0"
algorand-python = "^2.0.0"
algorand-python-testing = "^0.4.0"
pytest = "^8.3.3"

[tool.poetry.group.dev.dependencies]
algokit-client-generator = "^1.1.3"
puyapy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

